requirements.txt
Flask==2.3.2
paramiko==3.5.1
PyMySQL==1.1.1
pillow==11.1.0
python-dotenv==1.0.0
gunicorn